requirements.txt
# Pantry-to-Plate System Requirements
# Python 3.9+

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP/API
requests>=2.31.0

# Database (Optional - for SQL extraction)
psycopg2-binary>=2.9.0

# LLM Integration (Optional)
anthropic>=0.18.0

# Web Framework (Optional - for REST API)
flask>=3.0.0
flask-cors>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

<<<<<<< HEAD
# Utilities
python-dateutil>=2.8.2
=======
# ML Model Loading
scikit-learn>=1.3.0
joblib>=1.3.0
pytorch_tabnet
torch

# Utilities
python-dateutil>=2.8.2
groq
Pillow
python-dotenv
>>>>>>> fulltabnetver
